# docs/site/static/api/swagger.yaml
openapi: 3.0.3
info:
  title: LynxDB API
  version: 1.0.0
  description: |
    Splunk-power log analytics in a single binary. Zero dependencies. SPL2-compatible.

    ## Design Principles

    1. **One URL — one resource.** `POST /query` does a query. `POST /ingest` ingests.
    2. **Streaming-first.** Large-data endpoints support NDJSON via `Accept: application/x-ndjson`.
    3. **Errors are UX.** Every error carries `code`, `message`, and optionally `suggestion` and `docs_url`.
    4. **Single auth model.** Bearer token. Optional. Off by default.
    5. **Zero required headers.** `Content-Type` defaults to `application/json`. `curl` just works.

    ## Authentication

    Off by default. When enabled via `lynxdb server --auth`:

    ```
    Authorization: Bearer <token>
    ```

    ## Response Envelope

    ```json
    // Success — "data" present, "error" absent
    {"data": {...}, "meta": {"took_ms": 89}}

    // Error — "error" present, "data" absent
    {"error": {"code": "INVALID_QUERY", "message": "...", "suggestion": "..."}}
    ```

    No `"status": "ok"`. HTTP status codes carry that.

  contact:
    name: LynxDB
    url: https://lynxdb.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:3100/api/v1
    description: Local development
...
security:
  - BearerAuth: []
  - {}

# ──────────────────────────────────────────────
# PATHS
# ──────────────────────────────────────────────
paths:
...
  /histogram:
    get:
      operationId: histogram
      tags: [Live]
      summary: Time-bucketed event counts
      description: |
        Powers the timeline bar chart in the Web UI. Fires on every filter change — must be blazing fast.
        Separate from `/query` because it runs independently and concurrently.
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: SPL2 filter expression (search part only, before pipes)
          example: "level=error"
        - name: from
          in: query
          required: true
          schema:
            type: string
          example: "-1h"
        - name: to
          in: query
          schema:
            type: string
            default: "now"
        - name: buckets
          in: query
          schema:
            type: integer
            default: 60
          description: Target number of buckets. Server picks the best interval (1s, 5s, 1m, 5m, 1h).
      responses:
        '200':
          description: Histogram data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistogramResponse'
              example:
                data:
                  interval: "1m"
                  buckets:
                    - time: "2026-02-14T14:00:00Z"
                      count: 42
                    - time: "2026-02-14T14:01:00Z"
                      count: 87
                    - time: "2026-02-14T14:02:00Z"
                      count: 156
                  total: 8432
                meta:
                  took_ms: 12

  # ════════════════════════════════════════════
  # SCHEMA
  # ════════════════════════════════════════════
  /fields:
    get:
      operationId: listFields
      tags: [Schema]
      summary: List all fields with types and stats
      description: Essential for autocomplete and the Fields sidebar in the Web UI.
      parameters:
        - name: from
          in: query
          schema:
            type: string
          description: Restrict stats to time range start
          example: "-1h"
        - name: to
          in: query
          schema:
            type: string
          description: Restrict stats to time range end
        - name: prefix
          in: query
          schema:
            type: string
          description: Filter fields by name prefix (for autocomplete)
          example: "sta"
        - name: source
          in: query
          schema:
            type: string
          description: Fields seen from a specific source
          example: "nginx"
      responses:
        '200':
          description: Field list with stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldsResponse'
              example:
                data:
                  fields:
                    - name: "_timestamp"
                      type: datetime
                      count: 847000000
                      coverage: 1.0
                    - name: level
                      type: string
                      count: 847000000
                      coverage: 1.0
                      top_values:
                        - value: info
                          count: 612000000
                        - value: error
                          count: 142000000
                        - value: warn
                          count: 93000000
                    - name: status
                      type: integer
                      count: 423000000
                      coverage: 0.50
                      min: 200
                      max: 504
                      top_values:
                        - value: 200
                          count: 380000000
                        - value: 404
                          count: 21000000
                    - name: duration_ms
                      type: float
                      count: 423000000
                      coverage: 0.50
                      min: 0.1
                      max: 30001.0
                      avg: 145.3
                      p50: 42.0
                      p99: 3200.0
  /fields/{name}/values:
    get:
      operationId: fieldValues
      tags: [Schema]
      summary: Top values for a specific field
      description: Powers "Quick Stats" sidebar and value autocomplete in the search bar.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          example: status
        - name: from
          in: query
          schema:
            type: string
          example: "-1h"
        - name: to
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        '200':
          description: Top field values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldValuesResponse'
              example:
                data:
                  field: status
                  type: integer
                  values:
                    - value: 200
                      count: 89421
                      percent: 72.3
                    - value: 404
                      count: 18234
                      percent: 14.7
                    - value: 500
                      count: 9123
                      percent: 7.4
                    - value: 502
                      count: 4521
                      percent: 3.7
                    - value: 504
                      count: 2401
                      percent: 1.9
                  unique_count: 14
                  total_count: 123700
                meta:
                  from: "2026-02-14T13:52:00Z"
                  to: "2026-02-14T14:52:00Z"
...
  /alerts/{id}/test:
    post:
      operationId: testAlert
      tags: [Alerts]
      summary: Test an alert without sending notifications
      description: |
        Executes the alert query and evaluates the condition, but does **not** send notifications.
        Returns the query result and which channels **would** have fired.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertTestResponse'
              example:
                data:
                  would_trigger: true
                  result:
                    errors: 247
                  channels_that_would_fire:
                    - type: webhook
                      name: "Slack Ops"
                      status: reachable
                    - type: telegram
                      name: "SRE Chat"
                      status: reachable
                    - type: pagerduty
                      name: "P1 Escalation"
                      status: reachable
                  message: "Condition met: errors (247) > 100. Notifications NOT sent (test mode)."
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Optional. Off by default. Enable with `lynxdb server --auth`.
...
  schemas:
...
    Meta:
      type: object
      properties:
        took_ms:
          type: number
        scanned:
          type: integer
        query_id:
          type: string

    # ── Ingest ──
...
    FieldsResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [fields]
          properties:
            fields:
              type: array
              items:
                type: object
                required: [name, type, count, coverage]
                properties:
                  name:
                    type: string
                  type:
                    type: string
                    enum: [string, integer, float, boolean, datetime]
                  count:
                    type: integer
                  coverage:
                    type: number
                    description: Fraction of events containing this field (0.0–1.0)
                  min:
                    type: number
                  max:
                    type: number
                  avg:
                    type: number
                  p50:
                    type: number
                  p99:
                    type: number
                  top_values:
                    type: array
                    items:
                      type: object
                      properties:
                        value: {}
                        count:
                          type: integer
    FieldValuesResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [field, type, values]
          properties:
            field:
              type: string
            type:
              type: string
            values:
              type: array
              items:
                type: object
                properties:
                  value: {}
                  count:
                    type: integer
                  percent:
                    type: number
            unique_count:
              type: integer
            total_count:
              type: integer
        meta:
          type: object
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time
...
    HistogramResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [interval, buckets, total]
          properties:
            interval:
              type: string
              example: "1m"
            buckets:
              type: array
              items:
                type: object
                required: [time, count]
                properties:
                  time:
                    type: string
                    format: date-time
                  count:
                    type: integer
            total:
              type: integer
        meta:
          $ref: '#/components/schemas/Meta'

    # ── Status ──
...
    AlertTestResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            would_trigger:
              type: boolean
            result:
              type: object
              additionalProperties: true
            channels_that_would_fire:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                  name:
                    type: string
                  status:
                    type: string
                    enum: [reachable, unreachable, error]
            message:
              type: string

    # ── Dashboards ──
...